+++ packages/semi-ui-vue/src/components/icons/index.tsx
@@ -1,5 +1,5 @@
-import Icon, { IconProps, IconSize } from '@kousum/semi-icons-vue';
-
-export { IconProps, IconSize };
+import Icon from '@kousum/semi-icons-vue';
+import type { IconProps, IconSize } from '@kousum/semi-icons-vue';
+export type { IconProps, IconSize };
 
 export default Icon;