+++ packages/semi-ui-vue/src/components/navigation/SubNavTransition.tsx
@@ -1,4 +1,4 @@
-import {Transition} from '@kousum/semi-animation-vue';
+import {Transition} from '../../../../semi-animation-vue/src/animation/index';
 // import {Transition} from '../../../../semi-animation-vue/src/animation/index';
 import * as PropTypes from '../PropTypes';
 import {Motion} from '@douyinfe/semi-foundation/utils/type';
@@ -27,7 +27,6 @@ export const vuePropsType = {
 const SubNavTransition = defineComponent<SubNavTransitionProps>((props, {}) => {
 
 
-  const {isCollapsed, maxHeight = 999} = props;
   // eslint-disable-next-line no-unused-vars
   const immediate = ref(false);
 
@@ -37,30 +36,32 @@ const SubNavTransition = defineComponent<SubNavTransitionProps>((props, {}) => {
 
   const slots = useSlots()
 
-  return ()=>(
-    <Transition
-      from={{maxHeight: 0, opacity: 0}}
-      enter={{
-        maxHeight: {val: maxHeight, easing: 'easeInQuad', duration: 250},
-        opacity: {val: 1, duration: 200, easing: 'cubic-bezier(0.5, -0.1, 1, 0.4)'},
-      }}
-      leave={{
-        maxHeight: {val: 0, easing: ease, duration: 250},
-        opacity: {
-          val: 0,
-          duration: isCollapsed ? 1 : 200, // Need to be fast and transparent when put away, otherwise there will be jumping
-          easing: 'cubic-bezier(0.5, -0.1, 1, 0.4)',
-        },
-      }}
-      immediate={immediate.value}
-    >
-      {(transitionStyle: { maxHeight: number; opacity: number }) => {
+  return ()=>{
 
-        // console.log(11,slots.default?.(formatStyle(transitionStyle)))
-        return slots.default?.(formatStyle(transitionStyle))
-      }}
-    </Transition>
-  );
+    const {isCollapsed, maxHeight = 999} = props;
+    return (
+      <Transition
+        from={{maxHeight: 0, opacity: 0}}
+        enter={{
+          maxHeight: {val: maxHeight, easing: 'easeInQuad', duration: 250},
+          opacity: {val: 1, duration: 200, easing: 'cubic-bezier(0.5, -0.1, 1, 0.4)'},
+        }}
+        leave={{
+          maxHeight: {val: 0, easing: ease, duration: 250},
+          opacity: {
+            val: 0,
+            duration: isCollapsed ? 1 : 200, // Need to be fast and transparent when put away, otherwise there will be jumping
+            easing: 'cubic-bezier(0.5, -0.1, 1, 0.4)',
+          },
+        }}
+        immediate={immediate.value}
+        children={typeof slots.default === 'function' ? (transitionStyle: { maxHeight: number; opacity: number }) => {
+          return slots.default?.(formatStyle(transitionStyle))
+        }:slots.default}
+      >
+      </Transition>
+    )
+  };
 })
 
 SubNavTransition.props = vuePropsType