+++ packages/semi-ui-vue/src/components/navigation/__test__/NavigationDemo.tsx
@@ -33,7 +33,7 @@ const Test = defineComponent<ExampleProps>((props, {slots}) => {
         footer={{
           collapseButton: true,
         }}
-        isCollapsed={isCollapsed.value}
+        // isCollapsed={isCollapsed.value}
         bodyStyle={{height: 320}}
         items={[
           {itemKey: 'user', text: '用户管理', icon: <IconUser/>},