+++ packages/semi-ui-vue/vite.config.ts
@@ -8,6 +8,30 @@ export default defineConfig({
   server: {
     port: 8888
   },
+  build:{
+    lib: {
+      entry: path.resolve(__dirname, 'src/components/index.ts'),
+      formats: ['es'],
+      name: 'SemiUiVue',
+      // the proper extensions will be added
+      fileName: 'semi-ui-vue'
+    },
+    rollupOptions: {
+      // 确保外部化处理那些你不想打包进库的依赖
+      external: [
+        'vue',
+        '@douyinfe/semi-foundation',
+        '@douyinfe/semi-theme-default',
+        '@douyinfe/semi-animation'
+      ],
+      output: {
+        // 在 UMD 构建模式下为这些外部化的依赖提供一个全局变量
+        globals: {
+          vue: 'Vue'
+        }
+      }
+    },
+  },
   css: {
     preprocessorOptions: {
       scss: {